requirements.txt
cherrypy
paho-mqtt
requests
influxdb-client
telepot